refactor(puzzle): split z-order out of pile-detection (closes #270)
`reorderGroupsAfterDrop` and its helper `rectFullyContains` had nothing
to do with pile detection — they're a post-drop z-order concern that
just happened to be colocated because both ran in the same pipeline.

Move them (with their tests) to a new `src/game/z-order.ts` module so
`pile-detection.ts` can stay focused on the suppress-or-not heuristic.
