+++ src/game/pile-detection.ts
@@ -54,14 +54,6 @@ export function rectsOverlap(a: BoundingRect, b: BoundingRect): boolean {
 }
 
 /**
- * Check whether rect A fully contains rect B.
- */
-export function rectFullyContains(a: BoundingRect, b: BoundingRect): boolean {
-    return a.minX <= b.minX && a.minY <= b.minY &&
-           a.maxX >= b.maxX && a.maxY >= b.maxY;
-}
-
-/**
  * Expand a bounding rect by a padding amount in all directions.
  */
 export function padRect(rect: BoundingRect, padding: number): BoundingRect {
@@ -167,58 +159,4 @@ export function shouldSuppressMerge(
     // sections where matching groups are expected to overlap.
     return nonMateOverlapCount >= PILE_OVERLAP_THRESHOLD &&
            nonMateOverlapCount > mateOverlapCount;
-}
-
-/**
- * Reorder groups after a drop to prevent smaller groups from being
- * hidden behind larger ones.
- *
- * For each dropped group, checks if it fully covers any smaller groups.
- * If so, raises those covered groups above the dropped group so smaller
- * pieces are never hidden behind larger ones.
- *
- * @param droppedGroupIds - IDs of the groups that were just dropped
- * @param state - Current game state
- * @param bringGroupToFront - Function to bring a group to front (z-order)
- */
-export function reorderGroupsAfterDrop(
-    droppedGroupIds: number[],
-    state: GameState,
-    bringGroupToFront: (groupId: number) => void,
-): void {
-    const groupsToRaise: { groupId: number; size: number }[] = [];
-
-    for (const droppedId of droppedGroupIds) {
-        const droppedGroup = state.groups.find(g => g.id === droppedId);
-        if (!droppedGroup) continue;
-
-        const droppedBounds = pileBounds(droppedGroup, state.pieces);
-
-        // Check all other groups that are smaller than the dropped group
-        for (const otherGroup of state.groups) {
-            if (otherGroup.id === droppedId) continue;
-            if (otherGroup.pieces.size >= droppedGroup.pieces.size) continue;
-
-            const otherBounds = pileBounds(otherGroup, state.pieces);
-
-            // If the dropped group fully covers this smaller group, mark it for raising
-            if (rectFullyContains(droppedBounds, otherBounds)) {
-                // Avoid duplicates - a group might be covered by multiple dropped groups
-                if (!groupsToRaise.some(g => g.groupId === otherGroup.id)) {
-                    groupsToRaise.push({
-                        groupId: otherGroup.id,
-                        size: otherGroup.pieces.size,
-                    });
-                }
-            }
-        }
-    }
-
-    // Sort by size descending (largest raised first, smallest ends up on top)
-    groupsToRaise.sort((a, b) => b.size - a.size);
-
-    // Bring each covered group to front
-    for (const { groupId } of groupsToRaise) {
-        bringGroupToFront(groupId);
-    }
 }
